fixes after CR
